+++ decidim-admin/lib/decidim/admin/test/factories.rb
@@ -1,0 +1,7 @@
+FactoryGirl.define do
+  factory :participatory_process_user_role, class: Decidim::Admin::ParticipatoryProcessUserRole do
+    user
+    participatory_process
+    role "admin"
+  end
+end
