+++ decidim-admin/spec/factories.rb
@@ -1,9 +1,2 @@
-require_relative "../../decidim-core/spec/factories"
-
-FactoryGirl.define do
-  factory :participatory_process_user_role, class: Decidim::Admin::ParticipatoryProcessUserRole do
-    user
-    participatory_process
-    role "admin"
-  end
-end
+require "decidim/core/test/factories"
+require "decidim/admin/test/factories"
