+++ decidim-api/spec/factories.rb
@@ -1,2 +1,2 @@
-require_relative "../../decidim-core/spec/factories"
-require_relative "../../decidim-comments/spec/factories"
+require_relative "decidim/core/test/factories"
+require_relative "decidim/comments/test/factories"
