+++ decidim-comments/lib/decidim/comments/test/factories.rb
@@ -1,0 +1,9 @@
+require_relative "../../decidim-core/spec/factories"
+
+FactoryGirl.define do
+  factory :comment, class: Decidim::Comments::Comment do
+    author { build(:user, organization: commentable.organization) }
+    commentable { build(:participatory_process) }
+    body { Faker::Lorem.paragraph }
+  end
+end
